LayoutTests/canvas/philip/tests/2d.shadow.gradient.transparent.1-expected.txt

< [index] >
2d.shadow.gradient.transparent.1
Shadows are not drawn for transparent gradient fills
References: 2d.shadow.render
Actual output:
Expected output:

Passed
